# src/main/java/com/example/bankcards/dto/UserCreationDto.java
package com.example.bankcards.dto;

import com.example.bankcards.entity.Role;
import jakarta.validation.constraints.Email;
import jakarta.validation.constraints.NotBlank;
import jakarta.validation.constraints.NotNull;
import jakarta.validation.constraints.Size;
import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@NoArgsConstructor
@AllArgsConstructor
public class UserCreationDto {
    @NotBlank
    @Size(max = 50)
    private String name;

    @NotBlank
    @Size(max = 50)
    private String lastName;

    @NotNull
    @Size(max = 50)
    private String surname;

    @NotNull
    private Role role;

    @NotBlank
    @Email
    @Size(max = 50)
    private String login;

    @NotBlank
    @Size(min = 8, max = 12)
    private String password;
}