# src/main/java/com/example/bankcards/dto/UserPasswordDto.java
package com.example.bankcards.dto;

import jakarta.validation.constraints.NotBlank;
import jakarta.validation.constraints.NotNull;
import jakarta.validation.constraints.Size;
import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@NoArgsConstructor
@AllArgsConstructor
public class UserPasswordDto {
    @NotNull
    private Long userId;

    @NotBlank
    @Size(min = 8, max = 12)
    private String password;
}