# src/main/java/com/example/bankcards/dto/auth/AuthenticationRequest.java
package com.example.bankcards.dto.auth;

import jakarta.validation.constraints.NotBlank;
import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@AllArgsConstructor
@NoArgsConstructor
public class AuthenticationRequest {
    @NotBlank
    private String login;
    @NotBlank
    private String password;
}